bear/requirements.txt
fastapi==0.75.0
numpy==1.21.0
Pillow==9.2.0
tensorflow==2.8.0
h5py==3.1.0
protobuf==3.20.*
python-multipart
